requirements.txt
duckdb==1.1.1
pymongo==4.10.1
python-dotenv==1.0.1